% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{addRGBValuesToColor}
\alias{addRGBValuesToColor}
\title{RGB values for given color hexcode
Get the color hexcodes with their RGB values into a dataframe}
\usage{
addRGBValuesToColor(datawithcolorcodes)
}
\arguments{
\item{datawithcolorcodes}{accepts dataframe}
}
\value{
tempDF data frame which have a rgb values of hex code with their color codes
}
\description{
RGB values for given color hexcode
Get the color hexcodes with their RGB values into a dataframe
}
