% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{addUnsoldItemsIntoSales}
\alias{addUnsoldItemsIntoSales}
\title{Add unsold items from soh data into sales data}
\usage{
addUnsoldItemsIntoSales(salesDF, sohDF, i)
}
\arguments{
\item{salesDF}{sales dataframe}

\item{sohDF}{soh dataframe}

\item{i}{is integer takes 1 & 3 for i to differeniate training and periods}
}
\value{
finally return a data frame by combining sales and soh.
}
\description{
Add unsold items from soh data into sales data
}
