% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{closeDbConnection}
\alias{closeDbConnection}
\title{Close DB connections}
\usage{
closeDbConnection()
}
\arguments{
\item{:}{there is no inputs for this function}
}
\value{
ntg to return, close the DB connections which are in open
}
\description{
Close DB connections
}
