% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{dataSavingIntoCsvFiles}
\alias{dataSavingIntoCsvFiles}
\title{Create csv files from given data frame
It takes a all stores data with all attributes and save the data into csv files by store/producucttype.}
\usage{
dataSavingIntoCsvFiles(finalDF)
}
\arguments{
\item{finalDF}{It takes dataframe as input}
}
\value{
create a csv in the selected path
}
\description{
Create csv files from given data frame
It takes a all stores data with all attributes and save the data into csv files by store/producucttype.
}
