% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{getPeriodNumbers}
\alias{getPeriodNumbers}
\title{Get the FY period number and year for given dates}
\usage{
getPeriodNumbers(startDate, endDate, periodType)
}
\arguments{
\item{startDate}{is a character / Date}

\item{endDate}{accepts character / Date}

\item{periodType}{may be week,month,year}
}
\value{
Return data frame with period numbers for the give date ranges with Year
}
\description{
It takes character values
}
