% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{getSohForStartingWeekOfDate}
\alias{getSohForStartingWeekOfDate}
\title{Get soh for the starting period for the given dates if minimum of period is not equal to 2}
\usage{
getSohForStartingWeekOfDate(soh_period, sales_with_period, minperiod, minYear)
}
\arguments{
\item{soh_period}{soh data for given date ranges by increasing the period number of data}

\item{sales_with_period}{sales data merge with period}

\item{minperiod}{as integer starting period other than 2}

\item{minYear}{accepts character data (i.e starting year of the data)}
}
\value{
soh is dataframe having soh for the starting period of given dates
}
\description{
Get soh for the starting period for the given dates if minimum of period is not equal to 2
}
