% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPipelineAndAlgorithmFunctions.R
\name{getTrainingTestDataFromRawInput}
\alias{getTrainingTestDataFromRawInput}
\title{Seperate the traing and test(prediction) data from raw input data}
\usage{
getTrainingTestDataFromRawInput(Normalizedata, Starting_Date, Ending_Date)
}
\arguments{
\item{Normalizedata}{accepts dataframe}

\item{Selected_Data}{return the dataframe for the given time period range}
}
\description{
Seperate the traing and test(prediction) data from raw input data
}
