% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{getWeekOneSoh}
\alias{getWeekOneSoh}
\title{Get week-1 soh using sales happened in that week and soh of next week}
\usage{
getWeekOneSoh(soh_period, sales_with_period)
}
\arguments{
\item{soh_period}{accepts data frame (soh with edited period numbers for given date ranges by increasing the period number of data}

\item{sales_with_period}{is a dataframe sales data merge with period}
}
\value{
soh is dataframe i.e soh for week1
}
\description{
Get week-1 soh using sales happened in that week and soh of next week
}
