% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPipelineAndAlgorithmFunctions.R
\name{mlPipeline}
\alias{mlPipeline}
\title{Get the one hot encoding data}
\usage{
mlPipeline(inputForMLPipeline)
}
\arguments{
\item{inputForMLPipeline}{takes dataframe}
}
\value{
MLOutputNN return dataframe by performing numerical encoding and one hot encoding on input data
}
\description{
Get the one hot encoding data
}
