% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPipelineAndAlgorithmFunctions.R
\name{nnAlgorithm}
\alias{nnAlgorithm}
\title{NN algorithm function}
\usage{
nnAlgorithm(Training_set, Prediction_set, hidden_layers, step_max,
  pipeline_output)
}
\arguments{
\item{Training_set}{is a data frame to train the network}

\item{Prediction_set}{is also a data frame to test the NN}

\item{hidden_layers}{is a vector to specify the number of nurons in the hidden layers}

\item{step_max}{maximum steps for training the neural network}

\item{pipeline_output}{is used to denormalize the data}
}
\value{
ref a dataframe with actual and predicted values for the sales
}
\description{
NN algorithm function
}
