% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPipelineAndAlgorithmFunctions.R
\name{rfAlgorithm}
\alias{rfAlgorithm}
\title{RF algorithm function}
\usage{
rfAlgorithm(Training_set, Prediction_set, Ntree, Mtry)
}
\arguments{
\item{Training_set}{data frame to train the network}

\item{Prediction_set}{also a dataframe to test the model}

\item{Ntree, Mtry}{are numerical values, to fit the RF model}
}
\value{
ref resultant dataframe of RF algorithm
}
\description{
RF algorithm function
}
