% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessFunctions.R
\name{addRGBValuesToColor}
\alias{addRGBValuesToColor}
\title{RGB values for given color hexcode
Get the color hexcodes with their RGB values into a dataframe}
\usage{
addRGBValuesToColor(datawithcolorcodes, colorNames)
}
\arguments{
\item{datawithcolorcodes}{accepts dataframe}

\item{colorNames}{is dataframe which is having the 3 columns (one is actual color(colname:Revised.color.code.Reqd) codes and colors names provided by client and final column is HEXCODES (colname: HexCode)).}
}
\value{
tempDF data frame which have a rgb values of hex code with their color codes
}
\description{
RGB values for given color hexcode
Get the color hexcodes with their RGB values into a dataframe
}
