% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPipelineAndAlgorithmFunctions.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize the one hot encoding data to generate input for NN}
\usage{
NormalizeData(MlOutputData)
}
\arguments{
\item{MlOutputData}{is a dataframe}
}
\value{
the dataframe after did the normalisation
}
\description{
Normalize the one hot encoding data to generate input for NN
}
